completed basic calculator functionality.
I'll probably do a couple of more commits. I want to make it look good and set it up to be loaded into other pages well. I should probably integrate react to use useState for the bools and make use of the Component structure to pull this into other websites.